+++ cmake/cudawrappers-helper.cmake
@@ -1,152 +1,80 @@
-# =============================================================================
-# cmake-format: off
-# This module enables embedding kernel or source files (e.g. .cu) into targets
-# by linking their binary representation as static libraries. It also inlines
-# any local includes (#include "...") in the source file.
-#
-# Functions:
-#   - get_local_includes: Recursively collect local includes.
-#   - inline_local_includes: Prepend headers and remove #include "..."
-#   - target_embed_source: Embed a source file as binary and link to a target.
-#
-# cmake-format: on
-# =============================================================================
+include("${CMAKE_CURRENT_LIST_DIR}/inline-common.cmake")
 
 # cmake-format: off
-# Return a list of absolute file names for all the local includes of the
-# input_file. Only files in the root_dir will be considered.
 #
-# Parameters:
-#   input_file: The file to scan for includes
-#   root_dir:   The root directory in which to search for include files
-#   out_var:    Output variable name to receive the list of include files (absolute paths)
-# cmake-format: on
-# get_local_includes
-function(get_local_includes input_file root_dir out_var)
-  file(READ "${input_file}" input_file_content)
-  set(include_regex "(^|\r?\n)(#include[ \t]*\"([^\"]+)\")")
-  string(REGEX MATCHALL "${include_regex}" includes "${input_file_content}")
-
-  set(include_files_local "")
-
-  foreach(include ${includes})
-    # Extract the filename from the include directive
-    string(REGEX REPLACE "${include_regex}" "\\3" include_name "${include}")
-    file(GLOB_RECURSE include_paths "${root_dir}/*/${include_name}")
-    if(include_paths)
-      list(SORT include_paths ORDER DESCENDING)
-      list(GET include_paths 0 include_path)
-      get_local_includes("${include_path}" "${root_dir}" recursive_includes)
-      list(APPEND include_files_local ${recursive_includes} "${include_path}")
-    else()
-      message(
-        WARNING "Could not find include: ${include_name} in ${input_file}"
-      )
-    endif()
-  endforeach()
-
-  list(REMOVE_DUPLICATES include_files_local)
-  set(${out_var}
-      "${include_files_local}"
-      PARENT_SCOPE
-  )
-endfunction()
-
-# cmake-format: off
-# Create a new file with inlined (prepended) headers and local #include "..."
-# lines removed from the input file.
+# Inline a C/C++ source file with local includes as a binary object using `ld -b binary`.
 #
-# Parameters:
-#   input_file:     Absolute path to main source file
-#   output_file:    File to write the resulting inlined source
-#   include_files:  List of absolute paths to the files to inline
-# cmake-format: on
-# inline_local_includes
-function(inline_local_includes input_file output_file include_files)
-  file(READ "${input_file}" input_file_content)
-  set(include_regex "(^|\r?\n)(#include[ \t]*\"([^\"]+)\")")
-  string(REGEX REPLACE "${include_regex}" "" input_file_content
-                       "${input_file_content}"
-  )
-
-  set(output_content "")
-  foreach(include_file ${include_files})
-    file(READ "${include_file}" include_file_content)
-    string(REGEX REPLACE "${include_regex}" "" include_file_content
-                         "${include_file_content}"
-    )
-    set(output_content "${output_content}\n${include_file_content}")
-  endforeach()
-
-  set(output_content "${output_content}\n${input_file_content}")
-  file(WRITE "${output_file}" "${output_content}")
-endfunction()
-
-# cmake-format: off
-# Embed a source file in a static library and link it to a target. It inlines
-# any local includes (#include "...")
+# Usage:
+#   target_embed_source(<target_name> <input_file>)
 #
-# Parameters:
-#   target:     CMake target to link to
-#   input_file: Path to the file to embed
+# Produces:
+#   - A generated inlined source file with all local includes
+#   - A binary object (.o) compiled from it
+#   - A header file exposing _start, _end, _size symbols and a helper function
+#   - A static library target named after the input file's basename
+#   - Links this static library to the provided target
+#
 # cmake-format: on
-# target_embed_source
-function(target_embed_source target input_file)
-  include(CMakeDetermineSystem)
+#
+# target_embed_source(<target_name> <input_file>)
+function(target_embed_source target_name input_file)
+  get_filename_component(input_name "${input_file}" NAME)
+  get_filename_component(input_basename "${input_file}" NAME_WE)
+  get_filename_component(input_path "${input_file}" REALPATH)
 
-  get_filename_component(name "${input_file}" NAME_WLE)
-  get_filename_component(input_file_absolute "${input_file}" REALPATH)
+  file(RELATIVE_PATH output_source_file "${PROJECT_SOURCE_DIR}" "${input_path}")
+  set(output_object_file "${output_source_file}.o")
 
-  string(REPLACE "${PROJECT_SOURCE_DIR}" "${CMAKE_BINARY_DIR}"
-                 input_file_inlined "${input_file_absolute}"
-  )
+  set(all_deps "")
+  set(processed_files "")
+  get_dependencies("${input_path}" all_deps processed_files)
 
-  get_filename_component(
-    input_file_inlined_dir "${input_file_inlined}" DIRECTORY
-  )
-  file(MAKE_DIRECTORY "${input_file_inlined_dir}")
-
-  get_local_includes(
-    "${input_file_absolute}" "${PROJECT_SOURCE_DIR}" include_files
-  )
-
-  if("${include_files}" STREQUAL "")
-    configure_file("${input_file_absolute}" "${input_file_inlined}" COPYONLY)
-  else()
-    inline_local_includes(
-      "${input_file_absolute}" "${input_file_inlined}" "${include_files}"
-    )
-  endif()
-
-  file(RELATIVE_PATH input_file_inlined_relative "${PROJECT_SOURCE_DIR}"
-       "${input_file_absolute}"
-  )
-
-  set(embed_object_file "${CMAKE_CURRENT_BINARY_DIR}/${name}.o")
-  set(embed_tool ld)
-  set(embed_tool_args
-      -r
-      -b
-      binary
-      -A
-      ${CMAKE_SYSTEM_PROCESSOR}
-      -o
-      "${embed_object_file}"
-      "${input_file_inlined_relative}"
+  add_custom_command(
+    OUTPUT "${output_source_file}"
+    COMMAND
+      ${CMAKE_COMMAND} -Dinput_file="${input_path}"
+      -Doutput_file="${output_source_file}" -Droot_dir="${PROJECT_SOURCE_DIR}"
+      -P "${CMAKE_CURRENT_FUNCTION_LIST_DIR}/inline-local-includes.cmake"
+    DEPENDS "${input_path}" ${all_deps}
+    COMMENT "Inlining all includes of ${input_file}"
   )
 
   add_custom_command(
-    OUTPUT "${embed_object_file}"
-    COMMAND ${embed_tool} ARGS ${embed_tool_args}
-    WORKING_DIRECTORY "${CMAKE_BINARY_DIR}"
-    DEPENDS "${input_file_absolute}" "${input_file_inlined}" ${include_files}
+    OUTPUT "${output_object_file}"
+    COMMAND ${CMAKE_LINKER} -r -b binary -A ${CMAKE_SYSTEM_PROCESSOR} -o
+            "${output_object_file}" "${output_source_file}"
+    DEPENDS "${output_source_file}"
     COMMENT "Creating object file for ${input_file}"
+    VERBATIM
   )
 
-  if(NOT TARGET ${name})
-    add_library(${name} STATIC "${embed_object_file}")
-    set_target_properties(${name} PROPERTIES LINKER_LANGUAGE CXX)
-  endif()
+  string(REPLACE "." "_" symbol_base "${output_source_file}")
+  string(REPLACE "/" "_" symbol_base "${symbol_base}")
 
-  target_link_libraries(${target} PRIVATE ${name})
+  string(TOUPPER "${input_basename}_H_" include_guard)
+  set(header_file "${output_object_file}.h")
+  set(header_file "${CMAKE_BINARY_DIR}/${header_file}")
+  set(header_content
+      "#ifndef ${include_guard}
+#define ${include_guard}
+
+extern const unsigned char _binary_${symbol_base}_start[];
+extern const unsigned char _binary_${symbol_base}_end[];
+extern const unsigned int  _binary_${symbol_base}_size;
+
+const std::string ${input_basename}_source = std::string(
+        reinterpret_cast<const char*>(_binary_${symbol_base}_start),
+        reinterpret_cast<const char*>(_binary_${symbol_base}_end));
+
+#endif // ${include_guard}
+"
+  )
+
+  file(WRITE "${header_file}.in" "${header_content}")
+  configure_file("${header_file}.in" "${header_file}" @ONLY)
+
+  add_library(${input_basename} STATIC "${output_object_file}")
+  set_target_properties(${input_basename} PROPERTIES LINKER_LANGUAGE CXX)
+  target_include_directories(${target_name} PUBLIC ${CMAKE_BINARY_DIR})
+  target_link_libraries(${target_name} PRIVATE ${input_basename})
 endfunction()